% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photos.R
\name{sort_photos}
\alias{sort_photos}
\title{Sort photos by date}
\usage{
sort_photos(from, to, subdir = NULL)
}
\arguments{
\item{from}{character. Path to directory that contains photo files.}

\item{to}{character. Path to destination directory to move photo files. Will
be created automatically if it doesn't exist.}

\item{subdir}{character (default: \code{NULL}). Optional format to create subdirectory
based on date photo was first created. Use the POSIX formatting described
in \code{\link{strptime}}. For example, \code{"\%Y \%m \%B"} is the 4-digit
year, 2-digit month, and full month name.}
}
\value{
Invisibly returns path to destination directory (input argument
  \code{to}) as an \code{\link[fs]{fs_path}} object.
}
\description{
For each photo file, you can move it, copy it, skip it, or delete it.
}
\examples{
\dontrun{

sort_photos("~/Dropbox/Camera Uploads/", "~/Dropbox/family-photos/", subdir = "\%Y \%m \%B")
}

}
